src/requirements.txt
cryptography
pycryptodome
PyQt5
pem